requirements.txt
torch
torchvision
numpy<2
tensorboard
pyyaml
